pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rs00"
version = "0.1.0"
description = "Robostride RS00 motor driver (USB-CAN AT serial, proprietary extended-frame protocol)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = ["pyserial==3.5"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

